chore: add agent registry and lifecycle modules
